feat: add middlewares to main app
This commit adds a new function, `add_middlewares`, to the `main.py` file. This function is responsible for including the `ProcessTime` middleware in the FastAPI application. The `ProcessTime` middleware measures the time taken to process each request and adds the processing time as a header in the response.

A new file, `timeit.py`, is also added in the `middlewares` directory. This file contains the implementation of the `ProcessTime` middleware.

These changes improve the functionality of the application by adding the ability to measure the processing time of requests.